pyproject.toml
[project]
name = "django-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.4",
    "django-ninja>=1.3.0",
    "python-decouple>=3.8",
]

[dependency-groups]
dev = [
    "pre-commit>=4.0.1",
    "ruff>=0.9.2",
]

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint]
select = [
    # pycodestyle
    "E4",
    "E7",
    "E9",
    "W",
    # pyflakes
    "F",
    # isort
    "I",
    # pyupgrade
    "UP",
    # flake8-bandit
    "S",
    # flake8-django
    "DJ",
    # mccabe
    "C90",
]
